9_special_pythagorean_number.py 50%
Sheesh, this one is turning into such a mess of code. I need to come up with a firm plan. The linear transformations *do* generate pythag triples correctly, I just need to come up with a way to loop the calculations and return sets at a time. That and stop when a solution is found, and don't get stuck on one single thread- each transformation generates three pythag triples.